refactor(core-banking): 금액을 BIGINT/Long 으로 통일 — 데이터 사전 규약에 맞춘다
앞선 커밋(V20)에서 수신계 금액을 NUMERIC(15,0) 으로 바꾸며 payment 스키마를 기준으로
삼았다. 기준이 틀렸다. 정본은 레포 전체 규약인 데이터 사전이다.

  docs/data_dictionary.md
    | 금액 | amount / amt  | BIGINT 원 단위 |
    | 잔액 | balance / bal | BIGINT        |

customer_ddl_design.md 도 annual_income_amount·capital_amount 를 BIGINT "원화 단위"로
규정하고, loan 은 BIGINT 61개(자바 Long 67필드)로 이미 지키고 있었다. 어긋난 것은
수신계와 결제계뿐이었다.

  deposit  NUMERIC(18,2) → BIGINT  (32개 컬럼 / 11개 테이블, V20 재작성)
  payment  DECIMAL(15,0) → BIGINT  (7개 컬럼 / 4개 테이블, V25)

V20 에서 hold_amount 가 빠져 있었다. 컬럼 목록을 CREATE TABLE 본문만 훑어 만들었는데
이 컬럼은 V17 이 ALTER TABLE ADD COLUMN 으로 넣은 것이라 잡히지 않았다. 32개로 바로잡았다.

금리는 그대로 BigDecimal 이다. NUMERIC(5,2)·NUMERIC(6,4) 는 이율·세율이라 소수가
본질이고, 데이터 사전도 금리는 bps 단위 INT 로 금액과 구분한다. 저장 금액은 Long,
비율이 개입하는 파생 계산(이자·추천 점수)은 BigDecimal 로 두고 경계에서 감싼다.

부수 효과:
  - BigDecimal 스케일 함정이 사라진다. 100000 과 100000.00 이 equals 로 다른 값이던
    문제 때문에 곳곳에서 compareTo 를 써야 했다.
  - 결제계 원장 검증(차변합 == 대변합)이 정수 비교가 된다.
  - Account.withdraw 두 오버로드가 중복하던 출금가능액 판정을 assertWithdrawable 로 모았다.

검증 중 테스트가 두 번 내 실수를 잡았다. 금액 리터럴을 일괄 변환하면서 금리 단언
isEqualByComparingTo("3.50") 을 isEqualTo(3L) 로 잘라먹었고(18곳), 금리 0 단언도
정수로 바꿨다. 소수부가 있는 변환만 골라 되돌렸다.

+++ services/core-banking/src/main/java/com/bank/deposit/client/dto/TransactionApprovalVerifyRequest.java
@@ -12,5 +12,5 @@ public record TransactionApprovalVerifyRequest(
         String approvalToken,
         String fromAccountNo,
         String toAccountNo,
-        BigDecimal amount
+        Long amount
 ) {}
