+++ services/core-banking/src/main/java/com/bank/deposit/service/AccountService.java
@@ -12,7 +12,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.math.BigDecimal;
 import java.time.Clock;
 import java.time.LocalDate;
 import java.util.List;
@@ -81,8 +80,8 @@ public class AccountService {
     }
 
     @Transactional
-    public Account updateLimits(Long id, BigDecimal dailyWithdrawLimit,
-                                Integer dailyWithdrawCountLimit, BigDecimal atmWithdrawLimit) {
+    public Account updateLimits(Long id, Long dailyWithdrawLimit,
+                                Integer dailyWithdrawCountLimit, Long atmWithdrawLimit) {
         Account account = findById(id);
         account.updateLimits(dailyWithdrawLimit, dailyWithdrawCountLimit, atmWithdrawLimit);
         return account;
