+++ services/core-banking/src/main/java/com/bank/payment/api/dto/InboundPaymentResponse.java
@@ -1,12 +1,11 @@
 package com.bank.payment.api.dto;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 public record InboundPaymentResponse(
         String paymentInstructionId,
         String transactionNo,
-        BigDecimal transferAmount,
+        Long transferAmount,
         String status,
         OffsetDateTime requestedAt,
         OffsetDateTime completedAt,
